requirements.txt
Django==2.1.5
gunicorn==19.9.0
oscn==0.0.0.32
python-decouple==3.1
twilio==6.24.0
django-phonenumber-field==2.2.0
phonenumbers==8.10.5
